Weather v0.1.0 Weather.CLI

Module to process cli interface of weather app.

Functions

get_items_list(items, len)

Returns list of string from items to display for user.

get_max_len_of_names(items)

Returns max length of names through the weather fields (items).

Examples

iex> items = %{first: %{name: "Test"}, second: %{name: "Longer"}}
iex> Weather.CLI.get_max_len_of_names(items)
6

get_result_list(data)

Returns list of all string including title to display for user.

main(argv)

Main function witch starts the whole programm to fetch and display weather data from NOAA for any allowed location.

parse_args(argv)

Parses args from command line.

parse_param(name, value, extra \\ "")

Parses param from xpath.

Examples

iex> Weather.CLI.parse_param("name", "value")
%{name: "name", value: "value", extra: ""}

iex> Weather.CLI.parse_param("name", "value", "extra")
%{name: "name", value: "value", extra: "extra"}      

parse_response(data)

Parse xml data from successful http response.

process(station_id)

Processes data flow from fetching xml data for single station_id to displaing it as pretty and human readable table.

row_join(map, len)

Returns string for row from item’s values.

Examples

iex> Weather.CLI.row_join(%{name: "", value: "value", extra: ""}, 5)
"       value"

iex> Weather.CLI.row_join(%{name: "name", value: "value", extra: ""}, 5)
" name: value"

iex> Weather.CLI.row_join(%{name: "name", value: "value", extra: "extra"}, 
...> 5)
" name: value extra"